device: CC1352P: Add support for the 'other' variant
This brings back the 'other' board variant that is not a Launchpad
device.

The only differences are as was described in preinclude before:

xoscCapArrayDelta = 0xFA and that RF Antenna switch pins are DIO_6 and
DIO_5.

Note, that by default, DIO_6 is assigned to the RED led in the original
firmware, and toggeling the RED LED would also toggle the RF switch
chip! This is here now rectified by removing the LED existance.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>