Update CCS and SDK
The new version now is built against a new version of the SDK. Lets
update the Container as well.

While here, also update the documentation to properly reference the
correct/matched versions.

Note that the newer version of the SDK now uses LF line endings, so we
need to convert our patch to match (though --ignore-space-change to git
apply is a nice work-around).

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>